Added bottom navigation for (new) MainActivity, filled with ShoppingList fragments
