+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme" >
-        <activity android:name=".MainActivity" >
+        <activity android:name=".LoginActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -24,8 +24,8 @@
             android:label="TEST">
         </activity>
         <activity
-            android:name=".TemporaryActivity"
-            android:label="TEMP">
+            android:name=".MainActivity"
+            android:label="Main">
         </activity>
         <activity android:name="com.facebook.FacebookActivity"
             android:configChanges=
