+++ app/src/main/java/com/example/rummates/LoginActivity.java
@@ -1,0 +1,168 @@
+package com.example.rummates;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.bumptech.glide.request.RequestOptions;
+import com.example.rummates.signIn.SignIn;
+import com.facebook.AccessToken;
+import com.facebook.AccessTokenTracker;
+import com.facebook.CallbackManager;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
+import com.facebook.GraphRequest;
+import com.facebook.GraphResponse;
+import com.facebook.login.LoginResult;
+import com.facebook.login.widget.LoginButton;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.util.Arrays;
+
+//TODO:Rename to 'LoginActivity'
+//TODO:Create TabLayoutActivtiy(Which will be our new main activity after logging in)
+public class LoginActivity extends AppCompatActivity {
+
+    private LoginButton loginButton;
+    private Button magicButton, registerButton, signInButton;
+    String nick,password;
+
+    private CallbackManager callbackManager;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        loginButton = findViewById(R.id.login_button);
+        magicButton = findViewById(R.id.magic_button);
+
+        registerButton = findViewById(R.id.register_button);
+        signInButton = findViewById(R.id.sign_in_button);
+
+
+
+        callbackManager = CallbackManager.Factory.create();
+        loginButton.setPermissions(Arrays.asList("email","public_profile"));
+        checkLoginStatus();
+        loginButton.registerCallback(callbackManager, new FacebookCallback<LoginResult>() {
+            @Override
+            public void onSuccess(LoginResult loginResult) {
+                magicButton.setEnabled(true);
+            }
+
+            @Override
+            public void onCancel() {
+
+            }
+
+            @Override
+            public void onError(FacebookException error) {
+
+            }
+
+        });
+        magicButton.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View v) {
+                startActivity(new Intent(LoginActivity.this, PostsActivity.class));
+            }
+        });
+
+        registerButton.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View v) {
+                startActivity(new Intent(getApplicationContext(), RegisterActivity.class));
+            }
+        });
+
+        signInButton.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View v) {
+
+                EditText nickField = (EditText) findViewById(R.id.sign_in_username);
+                EditText passwordField = (EditText) findViewById(R.id.sign_in_password);
+
+                nick = nickField.getText().toString();
+                password = passwordField.getText().toString();
+
+                SignIn signIn = new SignIn(nick, password);
+                String status = signIn.handleSignIn(nick,password);
+                Toast.makeText(getBaseContext(), status, Toast.LENGTH_LONG).show();
+                if(Integer.parseInt(status) == 201){
+                    startActivity(new Intent(LoginActivity.this, MainActivity.class));
+                    //TODO go to Michal's user page
+                }
+            }
+        });
+    }
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
+        callbackManager.onActivityResult(requestCode,resultCode,data);
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
+    AccessTokenTracker tokenTracker = new AccessTokenTracker() {
+        @Override
+        protected void onCurrentAccessTokenChanged(AccessToken oldAccessToken, AccessToken currentAccessToken)
+        {
+            if(currentAccessToken==null) {
+                //magicButton.setEnabled(false);
+                Toast.makeText(LoginActivity.this,"User Logged out",Toast.LENGTH_LONG).show();
+            }
+            else {
+                loadUserProfile(currentAccessToken);
+                startActivity(new Intent(LoginActivity.this, MainActivity.class));
+            }
+
+        }
+    };
+
+    private void loadUserProfile(AccessToken newAccessToken)
+    {
+        GraphRequest request = GraphRequest.newMeRequest(newAccessToken, new GraphRequest.GraphJSONObjectCallback() {
+            @Override
+            public void onCompleted(JSONObject object, GraphResponse response)
+            {
+                try {
+                    String first_name = object.getString("first_name");
+                    String last_name = object.getString("last_name");
+                    String email = object.getString("email");
+                    String id = object.getString("id");
+                    String image_url = "https://graph.facebook.com/"+id+ "/picture?type=normal";
+
+                    RequestOptions requestOptions = new RequestOptions();
+                    requestOptions.dontAnimate();
+
+                    //Glide.with(LoginActivity.this).load(image_url).into(circleImageView);
+
+
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            }
+
+        });
+
+        Bundle parameters = new Bundle();
+        parameters.putString("fields","first_name,last_name,email,id");
+        request.setParameters(parameters);
+        request.executeAsync();
+
+    }
+
+    private void checkLoginStatus()
+    {
+        if(AccessToken.getCurrentAccessToken()!=null)
+        {
+            loadUserProfile(AccessToken.getCurrentAccessToken());
+        }
+    }
+}
