+++ app/src/main/res/layout/activity_main.xml
@@ -1,79 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:orientation="vertical"
-    android:gravity="center"
-    >
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textStyle="bold"
-        android:gravity="center"
-        android:textSize="66sp"
-        android:layout_marginTop="20dp"
-        android:id="@+id/title"
-        android:text="Roomies"
-        />
-    <EditText
-        android:id="@+id/sign_in_username"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textStyle="bold"
-        android:gravity="center"
-        android:layout_marginTop="100dp"
-        android:hint="Username"
-        android:inputType="textPersonName"
-        />
-    <EditText
-        android:id="@+id/sign_in_password"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textStyle="bold"
-        android:gravity="center"
-        android:layout_marginTop="25dp"
-        android:hint="Password"
-        android:inputType="textPassword"
-        />
-    <com.facebook.login.widget.LoginButton
-        android:id="@+id/login_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="50dp"
-        />
-    <Button
-        android:id="@+id/sign_in_button"
-        android:layout_width="200dp"
-        android:layout_height="40dp"
-        android:text="Sign In!"
-        android:layout_marginTop="10dp"/>
+    tools:context=".MainActivity">
 
-    <LinearLayout
+    <FrameLayout
+        android:id="@+id/fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:orientation="horizontal"
-        android:gravity="center"
-        >
-    <Button
-        android:id="@+id/magic_button"
-        android:layout_width="120dp"
-        android:layout_height="60dp"
-        android:text="test button"
-        android:layout_marginLeft="20dp"
-        android:layout_marginTop="5dp"/>
+        android:layout_above="@+id/bottom_navigation"/>
+        <!--android:layout_above="@+id/ma_bottom_navigation"-->
+        <!--TODO:Add bottom navigation-->
 
-    <Button
-        android:id="@+id/register_button"
-        android:layout_width="120dp"
-        android:layout_height="60dp"
-        android:layout_marginLeft="20dp"
-        android:text="register"
-        android:layout_marginTop="5dp"/>
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:id="@+id/bottom_navigation"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:layout_alignParentBottom="true"
+        app:menu="@menu/bottom_navigation"
+        android:background="?android:attr/windowBackground"/>
 
-</LinearLayout>
-</LinearLayout>
+
+</RelativeLayout>
