+++ app/src/main/res/menu/bottom_navigation.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/bottom_nav1"
+        android:icon="@drawable/ic_launcher_background"
+        android:title="Shopping List"/>
+
+    <item
+        android:id="@+id/bottom_nav2"
+        android:icon="@drawable/ic_launcher_background"
+        android:title="Notes"/>
+
+    <item
+        android:id="@+id/bottom_nav3"
+        android:icon="@drawable/ic_launcher_background"
+        android:title="Chat"/>
+
+</menu>
